refactor(infra/sequelize): separate projections (queries) and repos (commands)
