+++ .sequelizerc
@@ -7,12 +7,6 @@ module.exports = {
     'sequelize',
     'migrations'
   ),
-  'models-path': path.resolve(
-    'src',
-    'infra',
-    'sequelize',
-    'models'
-  ),
   'seeders-path': path.resolve(
     'src',
     'infra',
