+++ src/infra/sequelize/eventStore/index.ts
@@ -1,1 +1,2 @@
-export { SequelizeEventStore } from './sequelizeEventStore'
+export * from './eventStore.model'
+export * from './sequelizeEventStore'
