--- src/infra/sequelize/file/index.ts
@@ -1,1 +1,0 @@
-export * from './fileRepo'
