+++ src/infra/sequelize/index.ts
@@ -1,15 +1,11 @@
-import { FileRepo } from './file'
-import { NotificationRepo, makeGetFailedNotifications } from './notification'
-import { makeGetUnnotifiedProjectsForPeriode } from './project'
-import { initProjectProjections } from './project/projections'
+import { initProjectProjections } from './projections/project'
 import models from './models'
 import { EventStore } from '../../modules/eventStore'
 import { SequelizeEventStore } from './eventStore'
 
-export const fileRepo = new FileRepo(models)
-export const notificationRepo = new NotificationRepo(models)
-export const getFailedNotifications = makeGetFailedNotifications(models)
-export const getUnnotifiedProjectsForPeriode = makeGetUnnotifiedProjectsForPeriode(models)
+export * from './repos'
+export * from './projections'
+
 export const sequelizeEventStore = new SequelizeEventStore(models)
 
 export const initProjections = (eventStore: EventStore) => {
