--- src/infra/sequelize/models/file.ts
@@ -1,43 +1,0 @@
-import { DataTypes } from 'sequelize'
-
-export default (sequelize) => {
-  const File = sequelize.define(
-    'file',
-    {
-      id: {
-        type: DataTypes.UUID,
-        allowNull: false,
-        primaryKey: true,
-      },
-      filename: {
-        type: DataTypes.STRING,
-        allowNull: false,
-      },
-      forProject: {
-        type: DataTypes.UUID,
-        allowNull: true,
-      },
-      createdBy: {
-        type: DataTypes.UUID,
-        allowNull: true,
-      },
-      designation: {
-        type: DataTypes.STRING,
-        allowNull: false,
-      },
-      storedAt: {
-        type: DataTypes.STRING,
-        allowNull: true,
-      },
-    },
-    {
-      timestamps: true,
-    }
-  )
-
-  File.associate = (models) => {
-    // Add belongsTo etc. statements here
-  }
-
-  return File
-}
