--- src/infra/sequelize/notification/index.ts
@@ -1,2 +1,0 @@
-export * from './getFailedNotifications'
-export * from './notificationRepo'
