--- src/infra/sequelize/project/index.ts
@@ -1,1 +1,0 @@
-export * from './getUnnotifiedProjectsForPeriode'
