--- src/infra/sequelize/project/projections/onProjectNotificationDateSet.ts
@@ -1,29 +1,0 @@
-import { ProjectNotificationDateSet, ProjectNotified } from '../../../../modules/project/events'
-
-export const onProjectNotificationDateSet = (models) => async (
-  event: ProjectNotificationDateSet | ProjectNotified
-) => {
-  const ProjectModel = models.Project
-  const projectInstance = await ProjectModel.findByPk(event.payload.projectId)
-
-  if (!projectInstance) {
-    console.log(
-      'Error: onProjectNotificationDateSet projection failed to retrieve project from db',
-      event
-    )
-    return
-  }
-
-  // update notifiedOn
-  projectInstance.notifiedOn = event.payload.notifiedOn
-
-  try {
-    await projectInstance.save()
-  } catch (e) {
-    console.log(
-      'Error: onProjectNotificationDateSet projection failed to update project',
-      event,
-      e.message
-    )
-  }
-}
