+++ src/infra/sequelize/projections/file/index.ts
@@ -1,0 +1,1 @@
+export * from './file.model'
