+++ src/infra/sequelize/projections/index.ts
@@ -1,0 +1,3 @@
+export * from './file'
+export * from './notification'
+export * from './project'
