+++ src/infra/sequelize/projections/notification/queries/index.ts
@@ -1,0 +1,4 @@
+import models from '../../../models'
+import { makeGetFailedNotifications } from './getFailedNotifications'
+
+export const getFailedNotifications = makeGetFailedNotifications(models)
