+++ src/infra/sequelize/projections/project/index.ts
@@ -1,0 +1,3 @@
+export * from './queries'
+export * from './updates'
+export * from './project.model'
