+++ src/infra/sequelize/projections/project/queries/index.ts
@@ -1,0 +1,4 @@
+import models from '../../../models'
+import { makeGetUnnotifiedProjectsForPeriode } from './getUnnotifiedProjectsForPeriode'
+
+export const getUnnotifiedProjectsForPeriode = makeGetUnnotifiedProjectsForPeriode(models)
