+++ src/infra/sequelize/repos/README.md
@@ -1,0 +1,4 @@
+Repositories are destined to load aggregates and persist any changes to them.
+
+As a rule of thumb, they should only be invoked for Commands (see CQRS).
+For Queries, use a projection instead.
